requirements.txt
langchain-community
langchain-core
langchain-openai
python-dotenv
chromadb
openai
sqlalchemy
sqlalchemy-utils
pyodbc
